+++ playground-nuxt/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@tresjs/cientos": "^4.0.3",
     "@tresjs/nuxt": "^3.0.7",
-    "@tresjs/post-processing": "^0.7.1",
+    "@tresjs/post-processing": "workspace:*",
     "@types/three": "^0.169.0"
   },
   "devDependencies": {
