+++ src/core/pmndrs/helper.ts
@@ -1,0 +1,4 @@
+import type { InjectionKey, ShallowRef } from 'vue'
+import type { EffectComposer } from 'postprocessing'
+
+export const effectComposerInjectionKey: InjectionKey<ShallowRef<EffectComposer | null>> = Symbol('effectComposer')
