+++ src/core/pmndrs/index.ts
@@ -1,32 +1,21 @@
-/* eslint-disable perfectionist/sort-named-exports */
+export { useEffectPmndrs } from './composables/useEffectPmndrs.js'
+export type {
+  BloomPmndrsProps,
+  DepthOfFieldPmndrsProps,
+  EffectComposerPmndrsProps,
+  GlitchPmndrsProps,
+  NoisePmndrsProps,
+  OutlinePmndrsProps,
+  PixelationPmndrsProps,
+  VignettePmndrsProps,
+} from './types.js'
 
-import Bloom, { type BloomProps } from './Bloom.vue'
-import { useEffect } from './composables/useEffect'
-import DepthOfField, { type DepthOfFieldProps } from './DepthOfField.vue'
-import EffectComposer, { type EffectComposerProps } from './EffectComposer.vue'
-import Glitch, { type GlitchProps } from './Glitch.vue'
-import Noise, { type NoiseProps } from './Noise.vue'
-import Outline, { type OutlineProps } from './Outline.vue'
-import Pixelation, { type PixelationProps } from './Pixelation.vue'
-import Vignette, { type VignetteProps } from './Vignette.vue'
-
-export {
-  Bloom,
-  DepthOfField,
-  EffectComposer,
-  Glitch,
-  Noise,
-  Outline,
-  Pixelation,
-  useEffect,
-  Vignette,
-
-  BloomProps,
-  DepthOfFieldProps,
-  EffectComposerProps,
-  GlitchProps,
-  NoiseProps,
-  OutlineProps,
-  PixelationProps,
-  VignetteProps,
-}
+// eslint-disable-next-line perfectionist/sort-exports
+export { default as BloomPmndrs } from './BloomPmndrs.vue'
+export { default as DepthOfFieldPmndrs } from './DepthOfFieldPmndrs.vue'
+export { default as EffectComposerPmndrs } from './EffectComposerPmndrs.vue'
+export { default as GlitchPmndrs } from './GlitchPmndrs.vue'
+export { default as NoisePmndrs } from './NoisePmndrs.vue'
+export { default as OutlinePmndrs } from './OutlinePmndrs.vue'
+export { default as PixelationPmndrs } from './PixelationPmndrs.vue'
+export { default as VignettePmndrs } from './VignettePmndrs.vue'
