+++ src/core/three/Output.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import { OutputPass } from 'three/examples/jsm/postprocessing/OutputPass.js'
-import { useEffect } from './composables/useEffect'
+import { useEffect } from './composables/useEffect.js'
 
 const { pass } = useEffect(() => new OutputPass())
